Switch to ulong for timestamps

+++ src/camera/uvc_camera_node.h
@@ -55,8 +55,9 @@ class UVCCameraNode {
  public:
   UVCCameraNode(const UVCCameraConfig& config);
   ~UVCCameraNode();
-  void RegisterCallback(const std::function<void(std::shared_ptr<JpegBuffer>,
-                                                 double timestamp)>& callback);
+  void RegisterCallback(
+      const std::function<void(std::shared_ptr<JpegBuffer>,
+                               unsigned long timestamp)>& callback);
   void Start();
   void CallBack(uvc_frame_t* frame);  // This should not be used publicly
 
@@ -67,7 +68,7 @@ class UVCCameraNode {
   uvc_device_handle_t* device_handle_;
   uvc_stream_ctrl_t ctrl_;
   std::vector<
-      std::function<void(std::shared_ptr<JpegBuffer>, double timestamp)>>
+      std::function<void(std::shared_ptr<JpegBuffer>, unsigned long timestamp)>>
       callbacks_;
   std::atomic<bool> start_ = false;
 };
